requirements.txt
fastapi
uvicorn
psycopg2-binary
schedule
pydantic
google-generativeai
groq
python-dotenv